Feature/fcs 195/vent report months form (#49)
Added new components to show and store data entered to the report months form.
The new form gets displayed after user has saved the vent report period. When the user saves data on the form and these are valid and complete, the vent report is marked as complete on the task-list.